+++ skyperious/gui.py
@@ -8,7 +8,7 @@ Released under the MIT License.
 
 @author      Erki Suurjaak
 @created     26.11.2011
-@modified    27.02.2021
+@modified    28.02.2021
 ------------------------------------------------------------------------------
 """
 import ast
@@ -3036,6 +3036,7 @@ class DatabasePage(wx.Panel):
         label_progress   = wx.StaticText(panel_sync2)
         edit_sync_status = wx.TextCtrl(panel_sync2, size=(-1, 50), style=wx.TE_MULTILINE)
         check_contacts   = wx.CheckBox(panel_sync2, label="Do not overwrite existing &contact information")
+        check_older      = wx.CheckBox(panel_sync2, label="Do not check &older chats for messages to sync")
         button_sync      = controls.NoteButton(panel_sync2, bmp=images.ButtonMergeLeftMulti.Bitmap)
         button_sync_sel  = controls.NoteButton(panel_sync2, bmp=images.ButtonMergeLeft.Bitmap)
         button_sync_stop = controls.NoteButton(panel_sync2, bmp=images.ButtonStop.Bitmap)
@@ -3080,7 +3081,9 @@ class DatabasePage(wx.Panel):
         edit_sync_status.SetEditable(False)
         check_contacts.ToolTip = "Do not overwrite contact information from online data " \
                                  "if contact already exists in database"
-        check_contacts.Enabled = False
+        check_older.ToolTip = "Do not check older chats for messages to sync, " \
+                              "may take a long while"
+        check_contacts.Enabled = check_older.Enabled = False
         ColourManager.Manage(button_sync,      "BackgroundColour", "BgColour")
         ColourManager.Manage(button_sync_sel,  "BackgroundColour", "BgColour")
         ColourManager.Manage(button_sync_stop, "BackgroundColour", "BgColour")
@@ -3103,6 +3106,7 @@ class DatabasePage(wx.Panel):
         self.Bind(wx.EVT_CHECKBOX,   self.on_change_ctrl_login, check_login)
         self.Bind(wx.EVT_CHECKBOX,   self.on_change_ctrl_login, check_sync)
         self.Bind(wx.EVT_CHECKBOX,   self.on_change_ctrl_login, check_contacts)
+        self.Bind(wx.EVT_CHECKBOX,   self.on_change_ctrl_login, check_older)
         self.Bind(wx.EVT_BUTTON,     self.on_live_login,        button_login)
         self.Bind(wx.EVT_BUTTON,     self.on_live_sync,         button_sync)
         self.Bind(wx.EVT_BUTTON,     self.on_live_sync_sel,     button_sync_sel)
@@ -3125,6 +3129,7 @@ class DatabasePage(wx.Panel):
         self.label_sync_progress = label_progress
         self.edit_sync_status    = edit_sync_status
         self.check_skip_contacts = check_contacts
+        self.check_skip_older    = check_older
         self.button_sync         = button_sync
         self.button_sync_sel     = button_sync_sel
         self.button_sync_stop    = button_sync_stop
@@ -3165,6 +3170,7 @@ class DatabasePage(wx.Panel):
         sizer_sync2.Add(label_progress, border=5, flag=wx.ALL | wx.ALIGN_CENTER_HORIZONTAL)
         sizer_sync2.Add(edit_sync_status, proportion=1, border=5, flag=wx.ALL | wx.GROW)
         sizer_sync2.Add(check_contacts, border=5, flag=wx.ALL | wx.GROW)
+        sizer_sync2.Add(check_older,    border=5, flag=wx.ALL | wx.GROW)
         sizer_sync2.Add(button_sync, border=5, flag=wx.ALL | wx.GROW)
         sizer_sync2.Add(button_sync_sel, border=5, flag=wx.ALL | wx.GROW)
         sizer_sync2.Add(button_sync_stop, border=5, flag=wx.ALL | wx.GROW)
@@ -3173,7 +3179,7 @@ class DatabasePage(wx.Panel):
 
         sizer.Add(splitter, border=5, proportion=1, flag=wx.ALL | wx.GROW)
         splitter.SplitVertically(panel1, panel2)
-        pos = self.Size[1] / 3
+        pos = self.Size[1] / 4
         splitter_sync.SplitHorizontally(panel_sync1, panel_sync2, sashPosition=pos)
 
 
@@ -3217,6 +3223,7 @@ class DatabasePage(wx.Panel):
         self.check_login_auto.Enable(self.check_login_store.Value)
         self.check_login_sync.Enable(self.check_login_auto .Value)
         self.check_skip_contacts.Value = opts.get("skip_contact_update", False)
+        self.check_skip_older.Value    = opts.get("skip_older_chats", False)
         if self.db.live.is_logged_in() and self.button_login.Enabled:
             # Probable auto-login during HTML export
             self.edit_login_status.Value = 'Logged in to Skype as "%s"' % self.db.username
@@ -3256,6 +3263,7 @@ class DatabasePage(wx.Panel):
         elif ctrl is self.check_login_auto:    name = "auto"
         elif ctrl is self.check_login_sync:    name = "sync"
         elif ctrl is self.check_skip_contacts: name = "skip_contact_update"
+        elif ctrl is self.check_skip_older:    name = "skip_older_chats"
         if not self.check_login_store.Value: self.check_login_auto.Value = False
         if not self.check_login_auto .Value: self.check_login_sync.Value = False
         self.check_login_auto.Enable(self.check_login_store.Value)
@@ -3278,6 +3286,8 @@ class DatabasePage(wx.Panel):
             conf.Login[self.db.filename].pop("sync",     None)
         if not self.check_skip_contacts.Value:
             conf.Login[self.db.filename].pop("skip_contact_update", None)
+        if not self.check_skip_older.Value:
+            conf.Login[self.db.filename].pop("skip_older_chats", None)
         if not conf.Login[self.db.filename]: conf.Login.pop(self.db.filename)
         util.run_once(conf.save)